+++ client/packages/portal-pages/src/pages/context-page/ContextPage.tsx
@@ -1,45 +1,13 @@
 import { useFrameworkCurrentContext } from '@equinor/portal-core';
-import styled from 'styled-components';
 import { StyledMain } from '../common-styles/Styles';
-
-import { Card } from '@equinor/eds-core-react';
-import { createGlobalStyle } from 'styled-components';
-
-export const StyledBackground = createGlobalStyle`
-    body {
-       background: #e3e3e3;
-    };
-`;
-
-export const StyledContextPageGrid = styled.div`
-  display: grid;
-  gap: 1rem;
-  padding: 2rem;
-
-  @media only screen and (min-width: 45rem) and (max-width: 80rem) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-  @media only screen and (min-width: 80rem) {
-    grid-template-columns: repeat(4, 1fr);
-  }
-`;
-
-export const StyledGridItem = styled(Card)<{
-  span?: number;
-  heightSpan?: number;
-}>`
-  @media only screen and (min-width: 45rem) {
-    grid-column: span ${({ span }) => span || 1};
-    grid-row: span ${({ heightSpan }) => heightSpan || 1};
-  }
-`;
+import {
+  StyledBackground,
+  StyledContextPageGrid,
+  StyledGridItem,
+} from './ContextPage.Styles';
 
 export const ContextPage = () => {
   const currentContext = useFrameworkCurrentContext();
-
-  // if (!currentContext?.id) {
-  //   return <Navigate to="/" />;
-  // }
 
   return (
     <StyledMain>
